Flash: Hide the flash window when opacity reaches 10%.
The way this work isn't ideal and it should be redesigned rather than
using a treshold value.

Anyhow, we can observe an opacity stagnating at around 0.03 and never reaching
the 0.01 threshold. This results in the flash window remaining visible on the screen.

When the screenshot taken is an area/window, that area remains visible as a dimmed
rectangle until gnome-screenshot is killed.

This commit prevents this from happening by setting a higher threshold.